VMC/FlightSoftware/pcc/requirements.txt
loguru==0.5.3
paho-mqtt==1.5.1
pyserial==3.5